Refactoring of context objects into the Context class.
This means that we now have a central composed object that's created
when the parsing starts threaded through passes all the way to code generation.

For the sake of consistency, the "top level" definitions behave exactly
like "local" definitons would, i.e. names can be re-bound or "shadowed".

We also introduced a `Marker` object that would check the consistency of
attribute combinations.